Adress comments 2.
