+++ include/aspect/parameters.h
@@ -326,12 +326,10 @@ namespace aspect
     double                         composition_solver_tolerance;
     unsigned int                   max_pre_newton_nonlinear_iterations;
     unsigned int                   max_newton_line_search_iterations;
-
-    /// temporary place it in parameters.h till a better place is found
-    double              newton_residual;
-    double              switch_initial_newton_residual;
-    double              newton_theta;
-    double              minimum_linear_stokes_solver_tolerance;
+    double              		   switch_initial_newton_residual;
+    double              		   minimum_linear_stokes_solver_tolerance;
+    double              		   newton_theta;
+    double              		   newton_residual;
     /**
      * @}
      */
