+++ source/simulator/assemblers/newton_stokes.cc
@@ -19,8 +19,6 @@
 */
 
 #include <aspect/simulator.h>
-#include <aspect/utilities.h>
-#include <aspect/assembly.h>
 #include <aspect/newton.h>
 #include <aspect/simulator_access.h>
 #include <aspect/utilities.h>
@@ -78,30 +76,23 @@ namespace aspect
 
           const SymmetricTensor<4, dim> &stress_strain_director = scratch
                                                                   .material_model_outputs.stress_strain_directors[q];
-          const bool use_tensor = (stress_strain_director
-                                   != dealii::identity_tensor<dim>());
+
+          AssertThrow(stress_strain_director == dealii::identity_tensor<dim>(),
+        		  ExcMessage ("Error: The newton method currently doesn't support anisotropic viscosities."));
 
           const double JxW = scratch.finite_element_values.JxW(q);
 
           // TODO: Find out why in this version of ASPECT adding the derivative to the preconditioning
-          //is way worse than the normal preconitioning
-          if (true)//theta == 0)
+          //is way worse than the normal preconditioning
+          if (theta == 0)
             {
               for (unsigned int i = 0; i < stokes_dofs_per_cell; ++i)
                 for (unsigned int j = 0; j < stokes_dofs_per_cell; ++j)
                   if (scratch.dof_component_indices[i] ==
                       scratch.dof_component_indices[j])
-                    data.local_matrix(i, j) += ((
-                                                  /*use_tensor ?
-                                                  eta * 2.0 * (scratch.grads_phi_u[i]
-                                                         * stress_strain_director
-                                                         * scratch.grads_phi_u[j]) :*/
-                                                  eta * (scratch.grads_phi_u[i]
-                                                         * scratch.grads_phi_u[j]))
-                                                + one_over_eta * pressure_scaling
-                                                * pressure_scaling
-                                                * (scratch.phi_p[i] * scratch
-                                                   .phi_p[j]))
+                    data.local_matrix(i, j) += ((eta * (scratch.grads_phi_u[i] * scratch.grads_phi_u[j]))
+                                                + one_over_eta * pressure_scaling * pressure_scaling
+                                                * (scratch.phi_p[i] * scratch.phi_p[j]))
                                                * JxW;
             }
           else
@@ -114,27 +105,15 @@ namespace aspect
               const SymmetricTensor<2,dim> viscosity_derivative_wrt_strain_rate = derivatives->viscosity_derivative_wrt_strain_rate[q];
               const SymmetricTensor<2,dim> strain_rate = scratch.material_model_inputs.strain_rate[q];
 
+              //todo: make this 0.9 into a global input parameter
               double alpha = Utilities::compute_spd_factor<dim>(eta, strain_rate, viscosity_derivative_wrt_strain_rate, 0.9);
-              /*const double denom = std::sqrt((strain_rate*strain_rate)*(viscosity_derivative_wrt_strain_rate*viscosity_derivative_wrt_strain_rate))-(strain_rate*viscosity_derivative_wrt_strain_rate);
-              if (denom == 0)
-                alpha = 1.0;
-              else
-                {
-                  alpha = (2.0*eta)/denom;
-                  if (alpha >= 1.0)
-                    alpha = 1.0;
-                  else
-                    alpha = std::max(0.0,0.9*alpha);
-                }*/
 
               for (unsigned int i = 0; i < stokes_dofs_per_cell; ++i)
                 for (unsigned int j = 0; j < stokes_dofs_per_cell; ++j)
                   if (scratch.dof_component_indices[i] ==
                       scratch.dof_component_indices[j])
                     {
-                      data.local_matrix(i, j) += ((
-                                                    eta * (scratch.grads_phi_u[i]
-                                                           * scratch.grads_phi_u[j]))
+                      data.local_matrix(i, j) += ((eta * (scratch.grads_phi_u[i] * scratch.grads_phi_u[j]))
                                                   + theta * alpha * (scratch.grads_phi_u[i] * (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[j]) * strain_rate
                                                                      + scratch.grads_phi_u[j] * (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[i]) * strain_rate)
                                                   + one_over_eta * pressure_scaling
@@ -187,7 +166,10 @@ namespace aspect
 
           const SymmetricTensor<4,dim> &stress_strain_director =
             scratch.material_model_outputs.stress_strain_directors[q];
-          const bool use_tensor = (stress_strain_director !=  dealii::identity_tensor<dim> ());
+
+          AssertThrow(stress_strain_director == dealii::identity_tensor<dim>(),
+        		  ExcMessage ("Error: The newton method currently doesn't support anisotropic viscosities."));
+
           const SymmetricTensor<2,dim> strain_rate = scratch.material_model_inputs.strain_rate[q];
           const double pressure = scratch.material_model_inputs.pressure[q];
           const double velocity_divergence = scratch.velocity_divergences[q];
@@ -211,10 +193,8 @@ namespace aspect
                   if (assemble_newton_stokes_matrix)
                     for (unsigned int j=0; j<stokes_dofs_per_cell; ++j)
                       {
-                        data.local_matrix(i,j) += ( /*(use_tensor ?
-                                                     eta * 2.0 * (scratch.grads_phi_u[i] * stress_strain_director * scratch.grads_phi_u[j])
-                                                     :*/
-                                                    eta * 2.0 * (scratch.grads_phi_u[i] * scratch.grads_phi_u[j])//)
+                        data.local_matrix(i,j) += (
+                                                    eta * 2.0 * (scratch.grads_phi_u[i] * scratch.grads_phi_u[j])
                                                     // assemble \nabla p as -(p, div v):
                                                     - (pressure_scaling *
                                                        scratch.div_phi_u[i] * scratch.phi_p[j])
@@ -240,18 +220,8 @@ namespace aspect
               const SymmetricTensor<2,dim> viscosity_derivative_wrt_strain_rate = derivatives->viscosity_derivative_wrt_strain_rate[q];
               const double viscosity_derivative_wrt_pressure = derivatives->viscosity_derivative_wrt_pressure[q];
 
+              //todo: make this 0.9 into a global input parameter
               double alpha  = Utilities::compute_spd_factor<dim>(eta, strain_rate, viscosity_derivative_wrt_strain_rate, 0.9);
-              /*const double denom = std::sqrt((strain_rate*strain_rate)*(viscosity_derivative_wrt_strain_rate*viscosity_derivative_wrt_strain_rate))-(strain_rate*viscosity_derivative_wrt_strain_rate);
-              if (denom == 0)
-                alpha = 1.0;
-              else
-                {
-                  alpha = (2.0*eta)/denom;
-                  if (alpha >= 1.0)
-                    alpha = 1.0;
-                  else
-                    alpha = std::max(0.0,0.9*alpha);
-                }*/
 
               for (unsigned int i=0; i<stokes_dofs_per_cell; ++i)
                 {
@@ -266,8 +236,6 @@ namespace aspect
                       {
                         data.local_matrix(i,j) += ( (eta * 2.0 * (scratch.grads_phi_u[i] * scratch.grads_phi_u[j]))
                                                     + theta * alpha * 2.0 * (scratch.grads_phi_u[i] * (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[j]) * strain_rate)
-                                                    //+ theta * alpha * (scratch.grads_phi_u[i] * (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[j]) * strain_rate
-                                                    //                + scratch.grads_phi_u[j] * (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[i]) * strain_rate)
                                                     + theta * pressure_scaling * scratch.grads_phi_u[i] * 2.0 * viscosity_derivative_wrt_pressure * scratch.phi_p[j] * strain_rate
                                                     // assemble \nabla p as -(p, div v):
                                                     - (pressure_scaling *
@@ -283,62 +251,15 @@ namespace aspect
                       }
 
                 }
-
-              /////// testing /////
-
-              //if(alpha != 1)
-              //  std::cout << "Set alpha: " << alpha << std::endl;
-              /*FullMatrix<double> A_T;
-              A_T.copy_transposed(data.local_matrix);
-              FullMatrix<double> x_loc = data.local_matrix;
-              x_loc.add(-1, A_T);
-              if(x_loc.frobenius_norm()/data.local_matrix.frobenius_norm() > 1e-8)
-                std::cout << "Not symetric!" << std::endl;*/
-              //else
-              //  std::cout << "Symmetric." << std::endl;
-
-
-              /*if (assemble_newton_stokes_matrix)
+#if debug
+              // Testing wheter the Jacobian is Symetric Positive Definite (SPD)
+              if (assemble_newton_stokes_matrix)
               {
                   SymmetricTensor<4,dim> H;
                   SymmetricTensor<4,dim> HT;
                   H = outer_product(strain_rate,viscosity_derivative_wrt_strain_rate);
                   HT = outer_product(viscosity_derivative_wrt_strain_rate,strain_rate);
                   SymmetricTensor<4,dim> SPD = 2 * eta * identity_tensor<dim>() + H + HT;
-
-                 // std::cout << "SPD = " << SPD << std::endl;
-
-
-              //                      bool testing_H = true;
-              //                      for (unsigned int sample = 0; sample < 1000; ++sample)
-              //                      {
-              //                          SymmetricTensor<2,dim> normal;
-              //
-              //                          for (unsigned int i=0; i<dim; ++i)
-              //                            for (unsigned int j=0; j<dim; ++j)
-              //                            {
-              //                              //std::cout << "Flag 2.1 = " << i << ":" << j << ", dim = " << dim << std::endl;
-              //                              normal[i][j] = Utilities::generate_normal_random_number (0, 1);
-              //                            }
-              //
-              //                          //std::cout << "Flag 3"<< std::endl;
-              //                          SymmetricTensor<2,dim> transposed = transpose(normal);
-              //
-              //                          double SPD_value = transposed * (SPD * normal);
-              //                          if(SPD_value < 0)
-              //                            testing_H = false;
-              //                       // std::cout << "SPD_value = " << SPD_value << std::endl;
-              //                      }
-              //                      if(testing_H == false)
-              //                      {
-              //                        std::cout << "H (SPD) is not spd!!! " << std::endl;
-              //                        std::cout << "{{" << SPD[0][0][0][0] << "," << SPD[0][0][1][1] << "," << SPD[0][0][0][1] << "," << SPD[0][0][1][0] << "}" << std::endl;
-              //                        std::cout << "{" << SPD[1][1][0][0] << "," << SPD[1][1][1][1] << "," << SPD[1][1][0][1] << "," << SPD[1][1][1][0] << "}" << std::endl;
-              //                        std::cout << "{" << SPD[0][1][0][0] << "," << SPD[0][1][1][1] << "," << SPD[0][1][0][1] << "," << SPD[0][1][1][0] << "}" << std::endl;
-              //                        std::cout << "{" << SPD[1][0][0][0] << "," << SPD[1][0][1][1] << "," << SPD[1][0][0][1] << "," << SPD[1][0][1][0] << "}}" << std::endl;
-              //                      }
-                  / *else
-                    std::cout << "H (SPD) are all spd :)" << std::endl;* /
 
 
               bool testing = true;
@@ -364,38 +285,23 @@ namespace aspect
                   }
                   std::cout << std::endl;
                   std::cout << std::setprecision(6) << std::endl;
+
+                  Assert(testing,ExcMessage ("Error: Assembly not SPD!."));
+
+                  // Testing whether all entries are finite.
                   for (unsigned int i=0; i<stokes_dofs_per_cell; ++i)
                   {
                     for (unsigned int j=0; j<stokes_dofs_per_cell; ++j)
                     {
-                      //std::cout << i << "," << j << "=" << data.local_matrix(i,j) << ", gpui = " << scratch.grads_phi_u[i] << ", eta = " << eta << ", gpuj = " << scratch.grads_phi_u[j] << ", deds = " << viscosity_derivative_wrt_strain_rate << ", ps = " << pressure_scaling << ", sr = " << strain_rate << ", " << std::sqrt(0.5*deviator(strain_rate)*deviator(strain_rate)) << std::endl;
-                      /*data.local_matrix(i,j) += ( // using tensor thing is complecating everything quite a bit, so removed it for now.
-                          scratch.grads_phi_u[i] * 2.0 * eta * scratch.grads_phi_u[j]
-                                                 + alpha * (scratch.grads_phi_u[i] * (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[j]) * strain_rate
-                                                     + scratch.grads_phi_u[j] * (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[i]) * strain_rate)
-                                                     - (pressure_scaling *
-                                                         scratch.div_phi_u[i] * scratch.phi_p[j])
-                                                         // finally the term -div(u). note the negative sign to make this
-                                                         // operator adjoint to the grad(p) term
-                                                         - (pressure_scaling *
-                                                             scratch.phi_p[i] * scratch.div_phi_u[j]))
-                                                             * scratch.finite_element_values.JxW(q);* /
                       Assert(dealii::numbers::is_finite(data.local_matrix(i,j)),ExcMessage ("Error: Assembly matrix is not finite."));
                     }
                   }
-
-                  Assert(testing,ExcMessage ("Error: Assembly not SSPD!."));
-                }else
-                {
-                  //if(abc != 0)
-                  //  std::cout << abc << " " << std::flush;
                 }
               }
               if(testing == false)
                 std::cout << std::endl;
-              }*/
-
-              ////// end testing //////
+              }
+#endif debug
 
             }
         }
@@ -437,12 +343,15 @@ namespace aspect
             }
 
           // Viscosity scalar
-          const double two_thirds = 2.0 / 2.0;//3.0;
+          const double two_thirds = 2.0 / 3.0;
           const double eta_two_thirds = scratch.material_model_outputs.viscosities[q] * two_thirds;
 
           const SymmetricTensor<4,dim> &stress_strain_director =
             scratch.material_model_outputs.stress_strain_directors[q];
-          const bool use_tensor = (stress_strain_director !=  dealii::identity_tensor<dim> ());
+
+          AssertThrow(stress_strain_director == dealii::identity_tensor<dim>(),
+        		  ExcMessage ("Error: The newton method currently doesn't support anisotropic viscosities."));
+
           const double velocity_divergence = scratch.velocity_divergences[q];
 
           const double JxW = scratch.finite_element_values.JxW(q);
@@ -453,12 +362,7 @@ namespace aspect
               for (unsigned int i=0; i<stokes_dofs_per_cell; ++i)
                 for (unsigned int j=0; j<stokes_dofs_per_cell; ++j)
                   {
-                    data.local_matrix(i,j) += (- /*(use_tensor ?
-                                                  eta_two_thirds * (scratch.div_phi_u[i] * trace(stress_strain_director * scratch.grads_phi_u[j]))
-                                                  :*/
-                                               eta_two_thirds * (scratch.div_phi_u[i] * scratch.div_phi_u[j])
-                                               /*)*/)
-                                              * JxW;
+                    data.local_matrix(i,j) += (- eta_two_thirds * (scratch.div_phi_u[i] * scratch.div_phi_u[j])) * JxW;
                   }
             }
           else
@@ -476,13 +380,7 @@ namespace aspect
               for (unsigned int i=0; i<stokes_dofs_per_cell; ++i)
                 for (unsigned int j=0; j<stokes_dofs_per_cell; ++j)
                   {
-                    //std::cout << "dv=" << theta * two_thirds * scratch.div_phi_u[i] * ( (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[j]) * velocity_divergence)
-                    //<< "dp=" << theta * two_thirds * (scratch.div_phi_u[i] * viscosity_derivative_wrt_pressure * scratch.phi_p[j]) * velocity_divergence << std::endl;
-                    data.local_matrix(i,j) += (- /*(use_tensor ?
-                                                  eta_two_thirds * (scratch.div_phi_u[i] * trace(stress_strain_director * scratch.grads_phi_u[j]))
-                                                  :*/
-                                               eta_two_thirds * (scratch.div_phi_u[i] * scratch.div_phi_u[j])
-                                               /*)*/
+                    data.local_matrix(i,j) += (-eta_two_thirds * (scratch.div_phi_u[i] * scratch.div_phi_u[j])
                                                - theta * two_thirds * scratch.div_phi_u[i] * ( (viscosity_derivative_wrt_strain_rate * scratch.grads_phi_u[j]) * velocity_divergence)
                                                - theta * two_thirds * (scratch.div_phi_u[i] * viscosity_derivative_wrt_pressure * scratch.phi_p[j]) * velocity_divergence
                                               )
